Unificacion de archivo UserService y UserController
